# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignCell API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/analytics/dashboard:
    get:
      operationId: getDashboardKpis
      tags: [analytics]
      summary: Get supervisor/admin dashboard KPIs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardKpis"
  /v1/analytics/conversion-trend:
    get:
      operationId: getConversionTrend
      tags: [analytics]
      summary: Get conversion rate trend over time
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Conversion trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendDataResponse"
  /v1/analytics/campaign-distribution:
    get:
      operationId: getCampaignDistribution
      tags: [analytics]
      summary: Campaign distribution by segment and type
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Distribution data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistributionDataResponse"
  /v1/analytics/expert-performance:
    get:
      operationId: getExpertPerformance
      tags: [analytics]
      summary: Expert performance metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Expert performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertPerformanceResponse"
  /v1/analytics/sla-compliance:
    get:
      operationId: getSlaCompliance
      tags: [analytics]
      summary: SLA compliance statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SLA compliance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaComplianceResponse"
  /v1/analytics/expert-kpis:
    get:
      operationId: getExpertKpis
      tags: [analytics]
      summary: Get KPIs for the logged-in campaign expert
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Expert KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertKpis"

  # ─── A/B EXPERIMENTS ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardKpis:
      type: object
      required: [totalCampaigns, activeCampaigns, conversionRate, aiAccuracy, slaCompliance, openCases, criticalCases, queuedCases]
      properties:
        totalCampaigns:
          type: integer
        activeCampaigns:
          type: integer
        conversionRate:
          type: number
        aiAccuracy:
          type: number
        slaCompliance:
          type: number
        openCases:
          type: integer
        criticalCases:
          type: integer
        queuedCases:
          type: integer
        avgConversionLift:
          type: number
        totalSubscribers:
          type: integer
    TrendPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: ["string", "null"]
    TrendDataResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    DistributionItem:
      type: object
      required: [name, value, count]
      properties:
        name:
          type: string
        value:
          type: number
        count:
          type: integer
    DistributionDataResponse:
      type: object
      required: [bySegment, byType]
      properties:
        bySegment:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        byType:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
    ExpertPerformanceEntry:
      type: object
      required: [expertId, name, completedCases, avgConversionLift, slaComplianceRate, avgResolutionHours]
      properties:
        expertId:
          type: string
        name:
          type: string
        completedCases:
          type: integer
        avgConversionLift:
          type: number
        slaComplianceRate:
          type: number
        avgResolutionHours:
          type: number
        points:
          type: integer
    ExpertPerformanceResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExpertPerformanceEntry"
    SlaComplianceResponse:
      type: object
      required: [compliance, breached, total, criticalCompliance]
      properties:
        compliance:
          type: number
        breached:
          type: integer
        total:
          type: integer
        criticalCompliance:
          type: number
    ExpertKpis:
      type: object
      required: [assignedCases, criticalCases, slaAtRisk, avgConversionLift, completedCases, aiAccuracy]
      properties:
        assignedCases:
          type: integer
        criticalCases:
          type: integer
        slaAtRisk:
          type: integer
        avgConversionLift:
          type: number
        completedCases:
          type: integer
        aiAccuracy:
          type: number
        totalPoints:
          type: integer

    # ── EXPERIMENT SCHEMAS ────────────────────────────────────────────────────
...