# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CampaignCell API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: string;
  value: number;
  /** @nullable */
  label?: string | null;
}